# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: KeyControl API
  version: "1.0.0"
  description: |
    # Overview
    Welcome to the official API documentation for the KeyControl gateway platform. This API provides the programmatic interface for securely managing, routing, and validating your API keys.

    ## Authentication
    This API is secured using Bearer authentication. With the exception of public authentication routes (`/api/auth/*`), all endpoints require a valid token passed in the `Authorization` header:

    `Authorization: Bearer <YOUR_TOKEN>`

    **Supported Token Types:**
    * **JWT:** Obtained via the standard user login flow.
    * **Master API Key:** A system-level key (prefixed with `mk-`) used for administrative access and automated integrations.
servers:
  - url: "{scheme}://{host}:{port}/api"
    description: Admin API
    variables:
      scheme:
        default: http
        enum: [http, https]
      host:
        default: localhost
      port:
        default: "3000"
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or Master API Key
      description: |
        Pass either:
        • A JWT returned by `POST /auth/login` (or `/auth/login/verify-2fa`).
        • The Master API Key (prefix `mk-`) generated via the Organization endpoints.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - METHOD_NOT_ALLOWED
            - PAYLOAD_TOO_LARGE
            - BAD_GATEWAY
            - GATEWAY_TIMEOUT
        details:
          description: Optional field-level validation details
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    SuccessFlag:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        api_key:
          type: string
        preset_id:
          type: integer
        preset_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api-keys/{id}:
    get:
      tags: [API Keys]
      operationId: getApiKey
      summary: Get a single API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API key details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [API Keys]
      operationId: updateApiKey
      summary: Update API key metadata or reassign preset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  maxLength: 500
                  nullable: true
                notes:
                  type: string
                  maxLength: 1000
                  nullable: true
                preset_id:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Updated API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        "400":
          description: Validation error or preset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [API Keys]
      operationId: deleteApiKey
      summary: Delete an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessFlag"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...